pyproject.toml
[build-system]
build-backend = "hatchling.build"
requires = [ "hatchling" ]

[project]
name = "umapjax"
version = "0.1.0"
description = "UMAP, but optimized with jax."
readme = "README.md"
license = { file = "LICENSE" }
maintainers = [ { name = "Adam Gayoso", email = "adamgayoso@gmail.com" } ]
authors = [ { name = "Adam Gayoso" } ]
requires-python = ">=3.11"
classifiers = [
  "Programming Language :: Python :: 3 :: Only",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
  "Programming Language :: Python :: 3.14",
]
dependencies = [ "chex", "jaxtyping", "mlx[cpu]>=0.30.4", "umap-learn" ]
# https://docs.pypi.org/project_metadata/#project-urls
urls.Homepage = "https://github.com/adamgayoso/umapjax"
urls.Source = "https://github.com/adamgayoso/umapjax"

[dependency-groups]
dev = [
  "pre-commit",
  "twine>=4.0.2",
  { include-group = "jax" },
  { include-group = "mlx" },
  { include-group = "torch" },
]
test = [
  "coverage>=7.10",
  "pytest",
  "pytest-cov",     # For VS Code’s coverage functionality
]
jax = [ "jax", "jaxlib" ]
mlx = [ "mlx[cpu]" ]
torch = [ "setuptools", "torch>=2", "torchvision" ]

[tool.hatch.build.targets.wheel]

packages = [ "src/umapjax" ]
[tool.hatch.envs.default]
installer = "uv"
dependency-groups = [ "dev" ]

# Test the lowest and highest supported Python versions with normal deps
[[tool.hatch.envs.hatch-test.matrix]]
deps = [ "stable" ]
python = [ "3.11", "3.14" ]

# Test the newest supported Python version also with pre-release deps
[[tool.hatch.envs.hatch-test.matrix]]
deps = [ "pre" ]
python = [ "3.14" ]

[tool.hatch.envs.hatch-test]
dependency-groups = [ "dev", "test" ]

[tool.hatch.envs.hatch-test.overrides]
# If the matrix variable `deps` is set to "pre",
# set the environment variable `UV_PRERELEASE` to "allow".
matrix.deps.env-vars = [
  { key = "UV_PRERELEASE", value = "allow", if = [
    "pre",
  ] },
]

[tool.ruff]
line-length = 120
src = [ "src" ]
extend-include = [ "*.ipynb" ]

format.docstring-code-format = true

lint.select = [
  "B",      # flake8-bugbear
  "BLE",    # flake8-blind-except
  "C4",     # flake8-comprehensions
  "D",      # pydocstyle
  "E",      # Error detected by Pycodestyle
  "F",      # Errors detected by Pyflakes
  "I",      # isort
  "RUF100", # Report unused noqa directives
  "TID",    # flake8-tidy-imports
  "UP",     # pyupgrade
  "W",      # Warning detected by Pycodestyle
]
lint.ignore = [
  "B008", # Errors from function calls in argument defaults. These are fine when the result is immutable.
  "D100", # Missing docstring in public module
  "D104", # Missing docstring in public package
  "D105", # __magic__ methods are often self-explanatory, allow missing docstrings
  "D107", # Missing docstring in __init__
  # Disable one in each pair of mutually incompatible rules
  "D203", # We don’t want a blank line before a class docstring
  "D213", # <> We want docstrings to start immediately after the opening triple quote
  "D400", # first line should end with a period [Bug: doesn’t work with single-line docstrings]
  "D401", # First line should be in imperative mood; try rephrasing
  "E501", # line too long -> we accept long comment lines; formatter gets rid of long code lines
  "E731", # Do not assign a lambda expression, use a def -> lambda expression assignments are convenient
  "E741", # allow I, O, l as variable names -> I is the identity matrix
  "F722", # syntax error in forward annotation, jaxtyping is not supported
]
lint.per-file-ignores."*/__init__.py" = [ "F401" ]
lint.per-file-ignores."docs/*" = [ "I" ]
lint.per-file-ignores."tests/*" = [ "D" ]
lint.pydocstyle.convention = "numpy"

[tool.pytest]
strict = true
testpaths = [ "tests" ]
addopts = [
  "--import-mode=importlib", # allow using test files with same name
]

[tool.coverage.run]
source = [ "umapjax" ]
patch = [ "subprocess" ]
omit = [ "**/test_*.py" ]

[tool.cruft]
skip = [
  "tests",
  "src/**/__init__.py",
  "src/**/basic.py",
  "docs/api.md",
  "docs/changelog.md",
  "docs/references.bib",
  "docs/references.md",
  "docs/notebooks/example.ipynb",
  "benchmark.ipynb",
]
